Ajout getOrderById pour retourner les infos de commande par l'id contenant les lots

+++ src/main/java/com/tracker/trackerapi/api/controller/LotController.java
@@ -84,7 +84,7 @@ public class LotController {
 
 
     @GetMapping("/order/lot/{order_id}")
-    public ResponseEntity<List<LotDto>> getLotsByOrder(@PathVariable int order_id) {
+    public ResponseEntity<List<LotDto>> getLotsByOrder(@PathVariable long order_id) {
         try {
             List<LotDto> lotDtoList = this.lotService.getLotsByOrder(order_id);
             return new ResponseEntity<>(lotDtoList, HttpStatus.OK);
